requirements.txt
streamlit==1.49.1
pandas==2.2.3
openpyxl==3.1.5
pillow==11.2.1
